aws-services/cdk/src/lambda-python/requirements.txt
requests==2.25.1
boto3
beautifulsoup4==4.9.3
arrow
pyswisseph==2.10.3.2